+++ src/app/main_tabs/heal_tab.rs
@@ -1,6 +1,6 @@
 use eframe::egui::Ui;
 
-use crate::{analyzer::*, custom_widgets::splitter::Splitter};
+use crate::{analyzer::*, app::settings::Settings, custom_widgets::splitter::Splitter};
 
 use super::{common::*, diagrams::*, tables::*};
 
@@ -27,8 +27,8 @@ impl HealTab {
         }
     }
 
-    pub fn update(&mut self, combat: &Combat) {
-        self.table = HealTable::new(combat, self.heal_group);
+    pub fn update(&mut self, settings: &Settings, combat: &Combat) {
+        self.table = HealTable::new(settings, combat, self.heal_group);
         self.main_diagrams = HealDiagrams::from_heal_groups(
             combat.players.values().map(self.heal_group),
             combat,
@@ -38,7 +38,7 @@ impl HealTab {
         self.selection_diagrams = None;
     }
 
-    pub fn show(&mut self, ui: &mut Ui) {
+    pub fn show(&mut self, settings: &Settings, ui: &mut Ui) {
         Splitter::horizontal()
             .initial_ratio(0.6)
             .ratio_bounds(0.1..=0.9)
@@ -52,7 +52,7 @@ impl HealTab {
                     );
                 });
 
-                self.show_diagrams(bottom_ui);
+                self.show_diagrams(settings, bottom_ui);
             });
     }
 
@@ -136,7 +136,7 @@ impl HealTab {
         }
     }
 
-    fn show_diagrams(&mut self, ui: &mut Ui) {
+    fn show_diagrams(&mut self, settings: &Settings, ui: &mut Ui) {
         ui.horizontal(|ui| {
             ui.selectable_value(
                 &mut self.active_diagram,
@@ -179,9 +179,9 @@ impl HealTab {
         }
 
         if let Some(selection_diagrams) = &mut self.selection_diagrams {
-            selection_diagrams.show(ui, self.active_diagram);
+            selection_diagrams.show(settings, ui, self.active_diagram);
         } else {
-            self.main_diagrams.show(ui, self.active_diagram);
+            self.main_diagrams.show(settings, ui, self.active_diagram);
         }
     }
 }