+++ src/app/main_tabs/tables/heal_table.rs
@@ -1,5 +1,8 @@
 use crate::{
-    analyzer::*, app::main_tabs::common::*, col, helpers::number_formatting::NumberFormatter,
+    analyzer::*,
+    app::{main_tabs::common::*, settings::Settings},
+    col,
+    helpers::number_formatting::NumberFormatter,
 };
 
 use super::metrics_table::*;
@@ -35,21 +38,28 @@ static COLUMNS: &[ColumnDescriptor<HealTablePartData>] = &[
             t.critical_percentage.show(r);
         },
     ),
-    col!("Ticks", |t| t.sort_by_desc(|p| p.ticks.all.count), |t, r| {
+    col!(
+        "Ticks",
+        |t| t.sort_by_desc(|p| p.ticks.all.count),
+        |t, r| {
             t.ticks.show(r);
         },
     ),
-    col!("Ticks / s",
+    col!(
+        "Ticks / s",
         "Ticks Per Second\nCalculated from the first action of the player to the last action in the log",
         |t| t.sort_by_option_f64_desc(|p| p.ticks_per_second.all.value),
         |t, r| {
             t.ticks_per_second.show(r);
         },
     ),
-    col!("Ticks %", |t| t.sort_by_option_f64_desc(|p| p.ticks_percentage.all.value), |t, r| {
-        t.ticks_percentage.show(r);
-    },
-),
+    col!(
+        "Ticks %",
+        |t| t.sort_by_option_f64_desc(|p| p.ticks_percentage.all.value),
+        |t, r| {
+            t.ticks_percentage.show(r);
+        },
+    ),
 ];
 
 pub struct HealTablePartData {
@@ -72,8 +82,18 @@ impl HealTable {
         Self::empty_base(COLUMNS)
     }
 
-    pub fn new(combat: &Combat, heal_group: impl FnMut(&Player) -> &HealGroup) -> Self {
-        Self::new_base(COLUMNS, combat, heal_group, HealTablePartData::new)
+    pub fn new(
+        settings: &Settings,
+        combat: &Combat,
+        heal_group: impl FnMut(&Player) -> &HealGroup,
+    ) -> Self {
+        Self::new_base(
+            settings,
+            COLUMNS,
+            combat,
+            heal_group,
+            HealTablePartData::new,
+        )
     }
 }
 
@@ -84,26 +104,48 @@ impl HealTablePart {
 }
 
 impl HealTablePartData {
-    fn new(group: &HealGroup, combat: &Combat, number_formatter: &mut NumberFormatter) -> Self {
+    fn new(
+        settings: &Settings,
+        group: &HealGroup,
+        combat: &Combat,
+        number_formatter: &mut NumberFormatter,
+    ) -> Self {
+        let more_decimals = settings.general.more_decimals;
         Self {
-            total_heal: ShieldAndHullTextValue::new(&group.total_heal, 2, number_formatter),
-            hps: ShieldAndHullTextValue::new(&group.hps, 2, number_formatter),
-            heal_percentage: ShieldAndHullTextValue::option(
-                &group.heal_percentage,
-                3,
+            total_heal: ShieldAndHullTextValue::new(
+                &group.total_heal,
+                if more_decimals { 2 } else { 0 },
                 number_formatter,
             ),
-            average_heal: ShieldAndHullTextValue::option(&group.average_heal, 2, number_formatter),
-            critical_percentage: TextValue::option(group.critical_percentage, 3, number_formatter),
+            hps: ShieldAndHullTextValue::new(
+                &group.hps,
+                if more_decimals { 2 } else { 0 },
+                number_formatter,
+            ),
+            heal_percentage: ShieldAndHullTextValue::option(
+                &group.heal_percentage,
+                if more_decimals { 3 } else { 2 },
+                number_formatter,
+            ),
+            average_heal: ShieldAndHullTextValue::option(
+                &group.average_heal,
+                if more_decimals { 2 } else { 0 },
+                number_formatter,
+            ),
+            critical_percentage: TextValue::option(
+                group.critical_percentage,
+                if more_decimals { 3 } else { 2 },
+                number_formatter,
+            ),
             ticks: ShieldAndHullTextCount::new(&group.heal_metrics.ticks),
             ticks_per_second: ShieldAndHullTextValue::new(
                 &group.ticks_per_second,
-                3,
+                if more_decimals { 3 } else { 1 },
                 number_formatter,
             ),
             ticks_percentage: ShieldAndHullTextValue::option(
                 &group.ticks_percentage,
-                3,
+                if more_decimals { 3 } else { 2 },
                 number_formatter,
             ),
             source_ticks: group.ticks.get(&combat.heal_ticks_manger).to_vec(),